static-automatic-ip.sh
If you frequently visit somewhere with a dodgy DHCP server (or one that doesn't hand out IPs to a guest network), you can use this to automatically set your IP and then revert when you leave. Edit the IP for your target address and change the Zenity prompt to fit your needs.